requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.49
botocore==1.35.49
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
jmespath==1.0.1
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
wheel==0.43.0
